sandbox-requirements.txt
dj-database-url
psycopg
django-csp
